refactor(server): one home for the request payload cap
The sandbox gateway and the webhook ingest surface each carried their own Content-Length filter,
with the same 411 and 413 answers written twice. Both now use core.web.PayloadSizeFilter; the
webhook filter keeps its prefix match and its per-provider refusal counters.
