+++ .changeset/one-home-for-the-payload-cap.md
@@ -1,0 +1,6 @@
+---
+---
+
+No release note: the webhook ingest surface and the sandbox gateway refuse exactly the requests they
+refused before, with the same status codes and the same per-provider counters. They now enforce the
+cap with one filter instead of two copies of it.
